refactor(chat): 🗑️ deprecate shared/message.ts
This is simpler.
